Name Change and Dead Space for Wheel Control
I changed the name of the robot file to 'main.c' so that people will
know the robot is using it . In addition, I added dead space to the
wheel control to prevent the boom from turning as the robot is driving.
(Note: Have not tested wheel dead space yet. I have only tested the boom
dead space)